# server/swagger.yaml
openapi: 3.0.0
info:
  title: ModernDoc API
  description: API endpoint for ModernDoc
servers:
  - url: http://localhost:8080
    description: ModernDoc API
...
paths:
...
  /file/get-by-id/{type}/{id}:
    get:
      tags:
        - File
      summary: Get file by id
      operationId: getFileById
      parameters:
        - name: type
          in: path
          description: file or folder
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: File's id
          required: true
          schema:
            type: string
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File retrieved
        '500':
          description: File retrieval failed
  /file/latest-files/{type}:
    get:
      tags:
        - File
      summary: Get latest files
      operationId: getLatestFiles
      parameters:
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      parameters:
        - name: type
          in: path
          description: file or folder
          required: true
          schema:
            type: string
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Files retrieved
        '500':
          description: Files retrieval failed
...
  /file/share/{type}/{fileId}:
    put:
      tags:
        - File
      summary: share file
      operationId: shareFile
      parameters:
        - name: type
          in: path
          description: file or folder
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          description: File's id
          required: true
          schema:
            type: string
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File shared
        '400':
          description: File share failed
  /file/unshare/{type}/{fileId}:
    put:
      tags:
        - File
      summary: unshare file
      operationId: unshareFile
      parameters:
        - name: type
          in: path
          description: file or folder
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          description: File's id
          required: true
          schema:
            type: string
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File unshared
        '400':
          description: File unshare failed
...